python_server/requirements.txt
flask
ariadne
couchdb
python-dotenv
flask-cors
requests 